fix new post adding, add image size restrictions, add errors modal

+++ client/src/components/Notification.jsx
@@ -4,24 +4,47 @@ import { useUser } from "./utilities/userContext";
 
 const Notification = ({ notifData }) => {
   const user = useUser();
+  const [deleted, setDeleted] = useState(false);
+
+  if (deleted) {
+    return <></>
+  }
 
   return (
     <div onClick={() => {
       axios.get(`http://localhost:8080/getPost?userId=${user.id}&postId=${notifData.postId}`).then(result => {
-        let data = JSON.parse(result.data.data);
-        
-        axios.get(`http://localhost:8080/getProfile?userId=${data.authorId}`).then(res => {
+        if (result.data.message === "Пост был удален") {
+          user.setErrorMessage("Пост был удален");
+          user.toggleError();
+          axios.delete(`http://localhost:8080/deleteNotification?notificationId=${notifData.id}`).then(deletionResult => {
+            console.log(deletionResult)
+            if (deletionResult.data.state === "Success") {
+              setDeleted(true);
+              user.toggleComments();
+              user.toggleNotifications();
+            } else {
+              user.setErrorMessage(deletionResult.data.message);
+              user.toggleError();
+            }
+          });
+        } else {
+          let data = JSON.parse(result.data.data);
+
+          axios.get(`http://localhost:8080/getProfile?userId=${data.authorId}`).then(res => {
           let authorAvatar = JSON.parse(res.data.data).imagePath;
           user.setSelectedPost({data, authorAvatar});
           axios.delete(`http://localhost:8080/deleteNotification?notificationId=${notifData.id}`).then(deletionResult => {
             if (deletionResult.data.state === "Success") {
+              setDeleted(true);
               user.toggleComments();
               user.toggleNotifications();
             } else {
-              console.log(deletionResult.data.message);
+              user.setErrorMessage(deletionResult.data.message);
+              user.toggleError();
             }
           });
         });
+        }
       });
     }}>
       Вас упомянули под постом
